Removed myokit.mxml. Moved 'split' into myokit.formats.xml, and 'html2ascii' into myokit.formats.html. Also moved the HTMLExporter and XMLExporter into those modules (formerly in formats.mathml). Closes #526.
