+++ myokit/formats/cellml/v1/_parser.py
@@ -17,7 +17,7 @@ import myokit.formats.mathml
 import myokit.formats.cellml as cellml
 import myokit.formats.cellml.v1
 
-from myokit.mxml import split
+from myokit.formats.xml import split
 
 
 def parse_file(path):
