+++ myokit/formats/mathml/__init__.py
@@ -1,5 +1,5 @@
 #
-# Provides MathML support
+# Provides MathML (version 2) support.
 #
 # This file is part of Myokit.
 # See http://myokit.org for copyright, sharing, and licensing details.
@@ -7,7 +7,6 @@
 from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 
-from ._exporter import XMLExporter, HTMLExporter
 from ._ewriter import MathMLExpressionWriter
 from ._parser import (   # noqa
     MathMLError,
@@ -21,18 +20,6 @@ NS_MATHML_2 = 'http://www.w3.org/1998/Math/MathML'
 
 # Importers
 # Exporters
-_exporters = {
-    'xml': XMLExporter,
-    'html': HTMLExporter,
-}
-
-
-def exporters():
-    """
-    Returns a dict of all exporters available in this module.
-    """
-    return dict(_exporters)
-
 
 # Expression writers
 _ewriters = {
