+++ myokit/tests/test_formats_exporters.py
@@ -180,6 +180,16 @@ class ExportTest(unittest.TestCase):
             self.assertRaisesRegex(
                 myokit.ExportError, 'file or link', e.runnable, dpath, m, p)
 
+    def test_completeness(self):
+        """
+        Test that all exporters have a test (so meta!).
+        """
+        methods = [x for x in dir(self) if x[:5] == 'test_']
+        for name in myokit.formats.exporters():
+            name = name.replace('-', '_')
+            name = 'test_' + name + '_exporter'
+            self.assertIn(name, methods)
+
     def test_ansic_exporter(self):
         self._test(myokit.formats.exporter('ansic'))
 
@@ -207,17 +217,14 @@ class ExportTest(unittest.TestCase):
     def test_easyml_exporter(self):
         self._test(myokit.formats.exporter('easyml'))
 
+    def test_html_exporter(self):
+        self._test(myokit.formats.exporter('html'))
+
     def test_latex_article_exporter(self):
         self._test(myokit.formats.exporter('latex-article'))
 
     def test_latex_poster_exporter(self):
         self._test(myokit.formats.exporter('latex-poster'))
-
-    def test_html_exporter(self):
-        self._test(myokit.formats.exporter('html'))
-
-    def test_xml_exporter(self):
-        self._test(myokit.formats.exporter('xml'))
 
     def test_matlab_exporter(self):
         self._test(myokit.formats.exporter('matlab'))
@@ -227,6 +234,22 @@ class ExportTest(unittest.TestCase):
 
     def test_opencl_rl_exporter(self):
         self._test(myokit.formats.exporter('opencl-rl'))
+
+    def test_opencl_exporter_errors(self):
+        # Checks the errors raised by the OpenCL exporter.
+        e = myokit.formats.exporter('opencl')
+
+        # Label membrane_potential must be set
+        m = myokit.load_model('example')
+        m.label('membrane_potential').set_label(None)
+        self.assertRaisesRegex(
+            ValueError, 'membrane_potential', self._test, e, m)
+
+        # Binding diffusion_current must be set
+        m = myokit.load_model('example')
+        m.binding('diffusion_current').set_binding(None)
+        self.assertRaisesRegex(
+            ValueError, 'diffusion_current', self._test, e, m)
 
     def test_python_exporter(self):
         self._test(myokit.formats.exporter('python'))
@@ -259,38 +282,15 @@ class ExportTest(unittest.TestCase):
             self.assertTrue(os.path.isdir(dpath))
             self.assertTrue(len(os.listdir(dpath)) > 0)
 
-    def test_opencl_exporter_errors(self):
-        # Checks the errors raised by the OpenCL exporter.
-        e = myokit.formats.exporter('opencl')
-
-        # Label membrane_potential must be set
-        m = myokit.load_model('example')
-        m.label('membrane_potential').set_label(None)
-        self.assertRaisesRegex(
-            ValueError, 'membrane_potential', self._test, e, m)
-
-        # Binding diffusion_current must be set
-        m = myokit.load_model('example')
-        m.binding('diffusion_current').set_binding(None)
-        self.assertRaisesRegex(
-            ValueError, 'diffusion_current', self._test, e, m)
-
-    def test_completeness(self):
-        """
-        Test that all exporters have a test (so meta!).
-        """
-        methods = [x for x in dir(self) if x[:5] == 'test_']
-        for name in myokit.formats.exporters():
-            name = name.replace('-', '_')
-            name = 'test_' + name + '_exporter'
-            self.assertIn(name, methods)
-
     def test_unknown_exporter(self):
         """
         Test the error handling for requesting an unknown exporter.
         """
         self.assertRaisesRegex(
             KeyError, 'Exporter not found', myokit.formats.exporter, 'elvish')
+
+    def test_xml_exporter(self):
+        self._test(myokit.formats.exporter('xml'))
 
 
 if __name__ == '__main__':
